Renamed integration to ExtraHop v2 	- updated references in playbooks and script 	- need to deprecate old ExtraHop integration
