+++ Integrations/ExtraHop/ExtraHop_description.md
@@ -1,1 +1,1 @@
-Visit the [ExtraHop + Demisto Setup Guide](https://www.extrahop.com/customers/community/bundles/extrahop/demisto-integration/) for detailed integration instructions.
+ExtraHop performs real-time stream analysis of the packets that carry data across a network.
