+++ Playbooks/playbook-ExtraHop_-_CVE-2019-0708_(BlueKeep).yml
@@ -90,7 +90,7 @@ tasks:
                 iscontext: true
               right:
                 value:
-                  simple: ExtraHop
+                  simple: ExtraHop v2
           - - operator: isEqualString
               left:
                 value:
@@ -350,10 +350,10 @@ tasks:
       version: -1
       name: ExtraHop Get Associated Transaction Records
       description: Query records associated with this detection from ExtraHop.
-      script: '|||extrahop-query-records'
+      script: 'ExtraHop v2|||extrahop-query-records'
       type: regular
       iscommand: true
-      brand: ""
+      brand: ExtraHop v2
     nexttasks:
       '#none#':
       - "22"
@@ -452,10 +452,10 @@ tasks:
       name: ExtraHop Get PCAP File
       description: Search for the specific packets associated with this detection
         in ExtraHop and return the pcap file.
-      script: '|||extrahop-search-packets'
+      script: 'ExtraHop v2|||extrahop-search-packets'
       type: regular
       iscommand: true
-      brand: ""
+      brand: ExtraHop v2
     nexttasks:
       '#none#':
       - "20"
