+++ Playbooks/playbook-ExtraHop_-_Default.yml
@@ -87,7 +87,7 @@ tasks:
                 iscontext: true
               right:
                 value:
-                  simple: ExtraHop
+                  simple: ExtraHop v2
           - - operator: isEqualString
               left:
                 value:
