+++ Playbooks/playbook-ExtraHop_-_Get_Peers_by_Host.yml
@@ -66,10 +66,10 @@ tasks:
       name: ExtraHop Get Peers
       description: Get the list of all of the peers that communicated with the given
         host in the specified time range.
-      script: '|||extrahop-get-peers'
+      script: 'ExtraHop v2|||extrahop-get-peers'
       type: regular
       iscommand: true
-      brand: ""
+      brand: ExtraHop v2
     nexttasks:
       '#none#':
       - "2"
@@ -129,10 +129,10 @@ tasks:
       name: ExtraHop Device Search
       description: Search for the device within ExtraHop using the provided name,
         IP address, and MAC address.
-      script: '|||extrahop-device-search'
+      script: 'ExtraHop v2|||extrahop-device-search'
       type: regular
       iscommand: true
-      brand: ""
+      brand: ExtraHop v2
     nexttasks:
       '#none#':
       - "11"
@@ -287,7 +287,7 @@ tasks:
                 iscontext: true
               right:
                 value:
-                  simple: ExtraHop
+                  simple: ExtraHop v2
           - - operator: isEqualString
               left:
                 value:
@@ -317,10 +317,10 @@ tasks:
       name: ExtraHop Get Activity Map
       description: Get a link to a live activity map for the given device in the specified
         time range.
-      script: '|||extrahop-get-activity-map'
+      script: 'ExtraHop v2|||extrahop-get-activity-map'
       type: regular
       iscommand: true
-      brand: ""
+      brand: ExtraHop v2
     nexttasks:
       '#none#':
       - "2"
