+++ Playbooks/playbook-ExtraHop_-_Ticket_Tracking.yml
@@ -42,10 +42,10 @@ tasks:
       description: Link the ExtraHop Detection to the corresponding Demisto Investigation.  This
         uses the ExtraHop ticket tracking functionality to properly display ticket
         status within ExtraHop.
-      script: '|||extrahop-track-ticket'
+      script: 'ExtraHop v2|||extrahop-track-ticket'
       type: regular
       iscommand: true
-      brand: ""
+      brand: ExtraHop v2
     nexttasks:
       '#none#':
       - "12"
@@ -442,10 +442,10 @@ tasks:
       description: Link the ExtraHop Detection to the corresponding Demisto Investigation.  This
         uses the ExtraHop ticket tracking functionality to properly display ticket
         status within ExtraHop.
-      script: '|||extrahop-track-ticket'
+      script: 'ExtraHop v2|||extrahop-track-ticket'
       type: regular
       iscommand: true
-      brand: ""
+      brand: ExtraHop v2
     nexttasks:
       '#none#':
       - "11"
@@ -687,7 +687,7 @@ tasks:
                 iscontext: true
               right:
                 value:
-                  simple: ExtraHop
+                  simple: ExtraHop v2
           - - operator: isEqualString
               left:
                 value:
